Move heading toolbar component so it can be resused
