+++ assets/js/components/heading-toolbar/heading-level-icon.js
