+++ assets/js/components/heading-toolbar/index.js
